work on chest
imported UV's mesh so dissolve shader will work. Tweaked anims. Current
bug that body of lock 2 doesn't render after reimport of mesh. I'll work
on that tomorrow, I know a work around but don;t have time tonight.
Added a rise up to reveal on the top level game object so the UI will
also rise up. Since it's imposible to just key frame the Y position, x
and z are now also key framed. We'll need to either put the prefab into
another game object at the scene level in order to move it around in the
scene, or change the reveal anim to key frame the chest and the UI
independently.