chord changes happening in progression + progression cheats
We now how certain voices being substituted as things progress, per
Guy's audio design.  I also have the progression cheats setup so you can
easily jump right into having any # of candles lit.  When the app
closes, we keep track of the further progression and then put a star
next to it next time the app is launched.  This is so we can resume
easily if we have to restart due to a bug or crash.